+++ Caramel/src/Render/Framebuffer.h
@@ -11,13 +11,13 @@ namespace Caramel
 		bool Bind();
 
 		// TEMP
-		void ResizeFBO(float a_width, float a_height);
+		void ResizeFBO(unsigned int a_width, unsigned int a_height);
 
 	public:
 		unsigned int m_FBO;
 		unsigned int m_colourTex;
 		unsigned int m_depthRender;
-		float m_FBOWidth, m_FBOHeight;
+		unsigned int m_FBOWidth, m_FBOHeight;
 		bool m_isValid;
 	};
 }
